requirements.txt
# backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases[sqlite]
pydantic
python-dotenv
openai
